refactor(server): extract connection_queries.go
Move connection CRUD, NewDatastore, Close, and GetPool into
connection_queries.go. Extract scanConnectionListItem and
scanFullConnection helpers.

Part of #75.
